Struct wasmer_bin::commands::Run

pub struct Run {
    disable_cache: bool,
    path: PathBuf,
    backend: Backend,
    invoke: Option<String>,
    em_symbol_map: Option<PathBuf>,
    em_entrypoint: Option<String>,
    pre_opened_directories: Vec<PathBuf>,
    mapped_dirs: Vec<String>,
    env_vars: Vec<String>,
    loader: Option<LoaderName>,
    resume: Option<String>,
    optimized_backends: Vec<Backend>,
    track_state: bool,
    call_trace: bool,
    block_trace: bool,
    command_name: Option<String>,
    cache_key: Option<String>,
    backend_llvm_options: LLVMCLIOptions,
    features: PrestandardFeatures,
    generate_debug_info: bool,
    args: Vec<String>,
}

Fields
disable_cache: bool

Disable the cache

path: PathBuf

Input file

backend: Backend

Name of the backend to use (x86_64)

invoke: Option<String>

Invoke a specified function

em_symbol_map: Option<PathBuf>

Emscripten symbol map

em_entrypoint: Option<String>

Begin execution at the specified symbol

pre_opened_directories: Vec<PathBuf>

WASI pre-opened directory

mapped_dirs: Vec<String>

Map a host directory to a different location for the wasm module

env_vars: Vec<String>

Pass custom environment variables

loader: Option<LoaderName>

Custom code loader

resume: Option<String>

Path to previously saved instance image to resume.

optimized_backends: Vec<Backend>

Optimized backends for higher tiers.

track_state: bool

Whether or not state tracking should be disabled during compilation. State tracking is necessary for tier switching and backtracing.

call_trace: boolblock_trace: boolcommand_name: Option<String>

The command name is a string that will override the first argument passed to the wasm program. This is used in wapm to provide nicer output in help commands and error messages of the running wasm program

cache_key: Option<String>

A prehashed string, used to speed up start times by avoiding hashing the wasm module. If the specified hash is not found, Wasmer will hash the module as if no cache-key argument was passed.

backend_llvm_options: LLVMCLIOptionsfeatures: PrestandardFeaturesgenerate_debug_info: bool

Generate debug information for use in a debugger

args: Vec<String>

Application arguments

Methods

impl Run

pub fn execute(&mut self)

pub fn parse_args(
&self,
module: &Module,
fn_name: &str
) -> Result<Vec<Value>, String>

Used with the invoke argument